check sha256 is valid skip downloaded file
